Rename img-node to nodes
