+++ src/esa-utils/thumbnail.ts
@@ -3,7 +3,7 @@ import type { Post } from "./models";
 import rehypeParse from "rehype-parse";
 import { find } from "unist-util-find";
 import { optimizeImage } from "./optimize-image";
-import { isImg } from "./img-node";
+import { isImg } from "./nodes";
 
 export function getFirstImg(post: Post): Promise<string> | undefined {
   const hast = unified().use(rehypeParse).parse(post.body_html);
